redirect_app/requirements.txt
fastapi
starlette
uvicorn
google-cloud-pubsub